refactor: remove SearchInput
